Segurança melhorada em todas as pagina, adicionada verificação das variaveis do url, corrigido nomes dos ficheiros, eliminados ficheiros desnecessários, layout do ficheiro de impressão dos orçamentos concluido, sideBar responsivo feito

+++ pages/fichaTrabalhoDelete.php
@@ -15,6 +15,12 @@
     }
 
     $idWorksheet = $_GET['idWorksheet'];
+    $sql = "SELECT * FROM worksheet WHERE id = '$idWorksheet'";
+    $result = $con->query($sql);
+    if ($result->num_rows <= 0) {
+        header('Location: dashboard.php');
+        exit();
+    }
 
     $sql = "DELETE FROM worksheet WHERE id = $idWorksheet;";
     $result = $con->prepare($sql);
