+++ pages/orcamentoDelete.php
@@ -15,6 +15,13 @@
 
     $idBudget = $_GET['idBudget'];
 
+    $sql = "SELECT * FROM budget WHERE id = '$idBudget'";
+    $result = $con->query($sql);
+    if ($result->num_rows <= 0) {
+        header('Location: dashboard.php');
+        exit();
+    }
+
     $sql = "DELETE FROM budget WHERE id = $idBudget;";
     $result = $con->prepare($sql);
     $result->execute();
